chore(spin_sdk::pg): Extract pg module into its own file
